Building A Model for Predicting Protein Expression Profile From A MultiModal Data
The goal of this Project is to predict the protein expression profile of 25 proteins in 1000 test cells based on the gene expression of 639 genes in those cells. Our objective is to leverage advanced predictive modeling techniques to accurately predict protein expression profiles based on gene expression data. We are provided with a training dataset containing information on the abundance of RNA transcripts and protein expression for a set of cells. The challenge lies in predicting the protein expression profile for a separate set of test cells, where some protein expression data has been intentionally altered. Drawing inspiration from methods such as multivariate regression and dimension reduction, our aim is to reconstruct the protein expression information that has been modified in the test set, utilizing the relationships between gene expression and protein markers observed in the training data.